pyproject.toml
[tool.poetry]
name = "kyve-slashing-monitoring"
version = "0.1.0"
description = ""
authors = ["mrworks <ostroushchenko.sa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
requests = "^2.28.0"
loguru = "^0.6.0"
PyYAML = "^6.0"
fastapi-utils = "^0.2.1"
"infi.clickhouse-orm" = "^2.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
